Supabase + linearlite + PGlite + Electric
Supabaseified version of the `linearlite` demo

Illustrates:

- offline first using only pglite (no tanstack)
- auth

The auth proxy is implemented in typescript via a Supabase Edge
function.

The changes write server is implemented in typescript, converted
to a plv8 function (via plv8ify) and executed as a Postgres stored
procedure. Supabase has plv8 installed by default.

Many rough edges still but implemented many of the features of the
original version of electric.